simple/rdm_tagged_peek.c: Keep context in scope between FI_PEEK|FI_CLAIM and FI_CLAIM.
The API requires that the context passed to fi_recvmsg() for
FI_PEEK|FI_CLAIM be a struct fi_context and that the same fi_context
must be passed to FI_CLAIM. The prior code met the API in the sense
that the fi_context to FI_CLAIM would have the same address, but it
had gone out of scope since the FI_PEEK and the contents were
lost. Since some providers hide the pointer to the claimed entry in
the context, this can lead to segfaults.

Making the context global keeps the fix as small as possible and
is consistent with the way the tests are written.

Signed-off-by: John Byrne <john.l.byrne@hpe.com>