poll: Fix race condition that can incorrectly report failure
The fi_poll test attempts to check if a provider incorrectly reports
too many transmit or receive events from calling fi_poll().  However,
there is a race condition where a provider does the following:

    if (value < last_read_value) {
        last_read_value = value;
        report counter;
    }

The fi_poll test then adds this check:

    if (tx_seq == fi_cntr_read(cntr))
        tx_done = 1;

The issue is that the counter value can change between the provider
saving the last_read_value and the test reading it.  So when the test
adds a check:

    if (tx_done)
        report error;

when fi_poll() returns the counter again, the test fails.  However,
the provider is correctly returning the counter based on its last
read value.  The application must be able to handle fi_poll() returning
even if the counter value did not change from its perspective.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>